#600 | feat(masonry): integrate BrickView as higher order component
